pyproject.toml
[project]
name = "selective-revocation-replay"
version = "0.1.0"
description = "Deterministic runtime and evaluation harness for selective revocation and replay after persistent indirect prompt injection."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "matplotlib>=3.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
